Declare atol/strtol/strtoul family with their C99 return widths
stdlib.h declared atol, strtol, strtoll, strtoul, strtoull, _strtoi64
and _strtoui64 as returning `int`, on the assumption that c5 stores
every integer as 64 bits. The codegen narrows a libc return to the
declared type's width, so an `int` declaration truncated these
functions' results to their low 32 bits. Values that fit in 32 bits
masked the defect; a result >= 2^32 (strtoul/atol) or >= 2^31 (strtol)
lost its high half -- e.g. strtoul("4294967296") returned 0.

Declare the standard return types per C99 7.20.1.2 / 7.20.1.4: long,
long long, unsigned long, unsigned long long. These track each
platform's type widths, matching the libc ABI (long is 64-bit on LP64,
32-bit on LLP64; long long is 64-bit on both). The post-call narrowing
after these calls drops out, visible in the libc_int_arith snapshot.

Surfaced by Tcl binary.test binary-37.12 (GetFormatSpec count
overflow): a count of 4294967296 read back as 0, so the overflow guard
never tripped. gcc-built tclsh passes the test; badc-built failed it.

Regression: strtoul_64bit_return.c checks strtoul/strtol/atol/strtoull
return full 64-bit values past 2^32.

+++ headers/include/stdlib.h
@@ -163,11 +163,16 @@ char *calloc(int n, int size);
 char *realloc(char *ptr, int size);
 int free(char *ptr);
 int atoi(char *s);
-int atol(char *s);
+// C99 7.20.1.2: atol returns long. A libc return wider than the
+// declared type is truncated to that type, so declaring it `int`
+// would drop the high half on LP64.
+long atol(char *s);
 double atof(char *s);
-int strtol(char *s, char **endp, int base);
-int strtoll(char *s, char **endp, int base);
-int _strtoi64(char *s, char **endp, int base);
+// C99 7.20.1.4: strtol/strtoll return long / long long; the declared
+// width must match so the 64-bit return register is not narrowed.
+long strtol(char *s, char **endp, int base);
+long long strtoll(char *s, char **endp, int base);
+long long _strtoi64(char *s, char **endp, int base);
 double strtod(char *s, char **endp);
 // C99 7.20.1.3. c5 stores every floating literal in `f64`, so
 // the prototype declares the return as double; the binding above
@@ -249,12 +254,12 @@ int __cxa_atexit(int *handler, char *arg, char *dso);
 #else
 int atexit(int *handler);
 #endif
-int strtoul(char *s, char **endp, int base);
-// C99 7.20.1.4: `strtoull` parses an unsigned long long. The c5
-// dialect stores integers as 64-bit values regardless of width
-// so the prototype returns `int` (also 64-bit on stack).
-int strtoull(char *s, char **endp, int base);
-int _strtoui64(char *s, char **endp, int base);
+// C99 7.20.1.4: strtoul returns unsigned long, strtoull returns
+// unsigned long long. The declared width must match the libc return
+// type; an `int` declaration truncates the result to its low 32 bits.
+unsigned long strtoul(char *s, char **endp, int base);
+unsigned long long strtoull(char *s, char **endp, int base);
+unsigned long long _strtoui64(char *s, char **endp, int base);
 #ifdef _WIN32
 // msvcrt's `_spawn*` family takes a mode argument up front.
 // `P_NOWAIT` returns the child handle immediately; the other
